+++ src/main/java/com/casaglass/casaglass_backend/dto/CorteInventarioCompletoDTO.java
@@ -22,10 +22,10 @@ public class CorteInventarioCompletoDTO {
     private String observacion;
     
     // Inventario por sede
-    private Integer cantidadInsula;
-    private Integer cantidadCentro;
-    private Integer cantidadPatios;
-    private Integer cantidadTotal;
+    private Double cantidadInsula;
+    private Double cantidadCentro;
+    private Double cantidadPatios;
+    private Double cantidadTotal;
     
     // Precios heredados de Producto
     private Double precio1;
@@ -35,7 +35,7 @@ public class CorteInventarioCompletoDTO {
     // Constructor principal con cálculo automático de total
     public CorteInventarioCompletoDTO(Long id, String codigo, String nombre, String categoria, String tipo, String color,
                                      Double largoCm, String observacion,
-                                     Integer cantidadInsula, Integer cantidadCentro, Integer cantidadPatios,
+                                     Double cantidadInsula, Double cantidadCentro, Double cantidadPatios,
                                      Double precio1, Double precio2, Double precio3) {
         this.id = id;
         this.codigo = codigo;
@@ -45,9 +45,9 @@ public class CorteInventarioCompletoDTO {
         this.color = color;
         this.largoCm = largoCm;
         this.observacion = observacion;
-        this.cantidadInsula = cantidadInsula != null ? cantidadInsula : 0;
-        this.cantidadCentro = cantidadCentro != null ? cantidadCentro : 0;
-        this.cantidadPatios = cantidadPatios != null ? cantidadPatios : 0;
+        this.cantidadInsula = cantidadInsula != null ? cantidadInsula : 0.0;
+        this.cantidadCentro = cantidadCentro != null ? cantidadCentro : 0.0;
+        this.cantidadPatios = cantidadPatios != null ? cantidadPatios : 0.0;
         this.precio1 = precio1;
         this.precio2 = precio2;
         this.precio3 = precio3;
@@ -58,24 +58,24 @@ public class CorteInventarioCompletoDTO {
     
     // Método para recalcular el total cuando se modifiquen las cantidades
     public void recalcularTotal() {
-        this.cantidadTotal = (cantidadInsula != null ? cantidadInsula : 0) + 
-                            (cantidadCentro != null ? cantidadCentro : 0) + 
-                            (cantidadPatios != null ? cantidadPatios : 0);
+        this.cantidadTotal = (cantidadInsula != null ? cantidadInsula : 0.0) + 
+                            (cantidadCentro != null ? cantidadCentro : 0.0) + 
+                            (cantidadPatios != null ? cantidadPatios : 0.0);
     }
     
     // Setters que recalculan automáticamente
-    public void setCantidadInsula(Integer cantidadInsula) {
-        this.cantidadInsula = cantidadInsula != null ? cantidadInsula : 0;
+    public void setCantidadInsula(Double cantidadInsula) {
+        this.cantidadInsula = cantidadInsula != null ? cantidadInsula : 0.0;
         recalcularTotal();
     }
     
-    public void setCantidadCentro(Integer cantidadCentro) {
-        this.cantidadCentro = cantidadCentro != null ? cantidadCentro : 0;
+    public void setCantidadCentro(Double cantidadCentro) {
+        this.cantidadCentro = cantidadCentro != null ? cantidadCentro : 0.0;
         recalcularTotal();
     }
     
-    public void setCantidadPatios(Integer cantidadPatios) {
-        this.cantidadPatios = cantidadPatios != null ? cantidadPatios : 0;
+    public void setCantidadPatios(Double cantidadPatios) {
+        this.cantidadPatios = cantidadPatios != null ? cantidadPatios : 0.0;
         recalcularTotal();
     }
 }
